--- GameLibrary/Pages/Admin/Create.cshtml
@@ -1,55 +1,0 @@
-﻿@page
-@model CreateModel
-
-@{
-    ViewData["Title"] = "Create";
-}
-
-<h1>Create</h1>
-
-<h4>Game</h4>
-<hr />
-<div class="row">
-    <div class="col-md-4">
-        <form method="post">
-            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-                <label asp-for="Game.Title" class="control-label"></label>
-                <input asp-for="Game.Title" class="form-control" />
-                <span asp-validation-for="Game.Title" class="text-danger"></span>
-            </div>
-			<div class="form-group">
-				<label asp-for="Game.ImageUrl" class="control-label"></label>
-				<input asp-for="Game.ImageUrl" class="form-control" />
-				<img src="@Model.Game.ImageUrl" class="img-thumbnail" />
-				<span asp-validation-for="Game.ImageUrl" class="text-danger"></span>
-			</div>
-            <div class="form-group">
-                <label asp-for="Game.Genre" class="control-label"></label>
-                <input asp-for="Game.Genre" class="form-control" />
-                <span asp-validation-for="Game.Genre" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Game.ReleaseDate" class="control-label"></label>
-                <input asp-for="Game.ReleaseDate" class="form-control" />
-                <span asp-validation-for="Game.ReleaseDate" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Game.Description" class="control-label"></label>
-                <input asp-for="Game.Description" class="form-control" />
-                <span asp-validation-for="Game.Description" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
-            </div>
-        </form>
-    </div>
-</div>
-
-<div>
-    <a asp-page="Index">Back to List</a>
-</div>
-
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
