--- GameLibrary/Pages/Admin/Delete.cshtml
@@ -1,53 +1,0 @@
-﻿@page
-@model DeleteModel
-
-@{
-    ViewData["Title"] = "Delete";
-}
-
-<h1>Delete</h1>
-
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <h4>Game</h4>
-    <hr />
-    <dl class="row">
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.Title)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.Title)
-        </dd>
-		<dt class="col-sm-2">
-			@Html.DisplayNameFor(model => model.Game.ImageUrl)
-		</dt>
-		<dd class="col-sm-10">
-			<img src="@Model.Game.ImageUrl" class="img-thumbnail" />
-			@Html.DisplayFor(model => model.Game.ImageUrl)
-		</dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.Genre)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.Genre)
-        </dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.ReleaseDate)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.ReleaseDate)
-        </dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.Description)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.Description)
-        </dd>
-    </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Game.Id" />
-        <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-page="./Index">Back to List</a>
-    </form>
-</div>
