--- GameLibrary/Pages/Admin/Details.cshtml
@@ -1,50 +1,0 @@
-﻿@page
-@model DetailsModel
-
-@{
-    ViewData["Title"] = "Details";
-}
-
-<h1>Details</h1>
-
-<div>
-    <h4>Game</h4>
-    <hr />
-    <dl class="row">
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.Title)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.Title)
-        </dd>
-		<dt class="col-sm-2">
-			@Html.DisplayNameFor(model => model.Game.ImageUrl)
-		</dt>
-		<dd class="col-sm-10">
-			<img src="@Model.Game.ImageUrl" class="img-thumbnail" />
-			@Html.DisplayFor(model => model.Game.ImageUrl)
-		</dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.Genre)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.Genre)
-        </dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.ReleaseDate)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.ReleaseDate)
-        </dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Game.Description)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Game.Description)
-        </dd>
-    </dl>
-</div>
-<div>
-    <a asp-page="./Edit" asp-route-id="@Model.Game.Id">Edit</a> |
-    <a asp-page="./Index">Back to List</a>
-</div>
