+++ GameLibrary/Pages/Games/Details.cshtml.cs
@@ -1,0 +1,39 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
+using GameLibrary.Data;
+using GameLibrary.Models;
+
+namespace GameLibrary.Pages.Games;
+
+public class DetailsModel : PageModel
+{
+    private readonly ApplicationDbContext _context;
+
+    public DetailsModel(ApplicationDbContext context)
+    {
+        _context = context;
+    }
+
+    public Game? Game { get; set; }
+
+    public async Task<IActionResult> OnGetAsync(int? id)
+    {
+        if (id == null)
+        {
+            return NotFound();
+        }
+
+        Game = await _context.Games.FirstOrDefaultAsync(m => m.Id == id);
+
+        if (Game == null)
+        {
+            return NotFound();
+        }
+        return Page();
+    }
+}
