+++ GameLibrary/Pages/Games/Edit.cshtml
@@ -1,0 +1,46 @@
+﻿@page
+@model GameLibrary.Pages.Games.EditModel
+
+@{
+    ViewData["Title"] = "Edit";
+}
+
+<h1>Edit</h1>
+
+<h4>Game</h4>
+<hr />
+<div class="row">
+    <div class="col-md-4">
+        <form method="post">
+            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            <input type="hidden" asp-for="Game!.Id" />
+            <div class="form-group">
+                <label asp-for="Game!.Title" class="control-label"></label>
+                <input asp-for="Game!.Title" class="form-control" />
+                <span asp-validation-for="Game!.Title" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="Game!.Genre" class="control-label"></label>
+                <input asp-for="Game!.Genre" class="form-control" />
+                <span asp-validation-for="Game!.Genre" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="Game!.ReleaseDate" class="control-label"></label>
+                <input asp-for="Game!.ReleaseDate" class="form-control" />
+                <span asp-validation-for="Game!.ReleaseDate" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="Game!.Description" class="control-label"></label>
+                <input asp-for="Game!.Description" class="form-control" />
+                <span asp-validation-for="Game!.Description" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <input type="submit" value="Save" class="btn btn-primary" />
+            </div>
+        </form>
+    </div>
+</div>
+
+<div>
+    <a asp-page="./Index">Back to List</a>
+</div>
