+++ GameLibrary/Pages/Games/Edit.cshtml.cs
@@ -1,24 +1,15 @@
-// Copyright 2024 Web.Tech. Group17
-//
-// Licensed under the Apache License, Version 2.0 (the "License"):
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using GameLibrary.Data;
-using GameLibrary.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using GameLibrary.Data;
+using GameLibrary.Models;
 
-namespace GameLibrary.Pages.Admin;
+namespace GameLibrary.Pages.Games;
 
 public class EditModel : PageModel
 {
@@ -39,17 +30,17 @@ public class EditModel : PageModel
             return NotFound();
         }
 
-        var game = await _context.Games.FirstOrDefaultAsync(m => m.Id == id);
-        if (game == null)
+        Game = await _context.Games.FirstOrDefaultAsync(m => m.Id == id);
+
+        if (Game == null)
         {
             return NotFound();
         }
-        Game = game;
         return Page();
     }
 
     // To protect from overposting attacks, enable the specific properties you want to bind to.
-    // For more information, see https://aka.ms/RazorPagesCRUD.
+    // For more details, see https://aka.ms/RazorPagesCRUD.
     public async Task<IActionResult> OnPostAsync()
     {
         if (!ModelState.IsValid)
